move base_core_search into own file, re #10804
